LC409E

+++ cpp/longest-increasing-subsequence.cpp
@@ -1,3 +1,4 @@
+
 // https://leetcode.com/problems/longest-increasing-subsequence/
 
 class Solution {
@@ -7,8 +8,8 @@ public:
         if (nums.size() == 1) return 1;
         vector<int> dp(nums.size(), 1);
         
-        for (int i=1; i<nums.size(); ++i) {
-            for (int j=0; j<i; ++j) {
+        for (size_t i=1; i<nums.size(); ++i) {
+            for (size_t j=0; j<i; ++j) {
                 if (nums[i]>nums[j])
                     dp[i] = max(dp[i], dp[j]+1);
             }
